requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.2
altair==5.3.0
annotated-types==0.6.0
anyio==4.4.0
APScheduler==3.10.4
attrs==23.2.0
biliup==0.4.70
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.11.17
cf_clearance==0.31.0
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
decorator==4.4.2
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.36.1
ffmpy==0.3.1
frozenlist==1.4.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
iniconfig==2.0.0
isodate==0.6.1
Jinja2==3.1.4
jsengine==1.0.7.post1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
lxml==5.2.2
m3u8==5.1.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
multidict==6.0.5
mutagen==1.47.0
mysql-connector-python==9.0.0
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.4
pillow==10.2.0
playwright==1.40.0
pluggy==1.5.0
premailer==3.10.0
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pycares==4.4.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.9.1
pyee==11.0.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypng==0.20220715.0
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
pyquery==2.0.0
PySocks==1.7.1
pytest==8.2.2
pytest-base-url==2.1.0
pytest-playwright==0.5.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.2rc1
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
schedule==1.2.2
setuptools==70.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
stream-gears==0.2.0
streamlink==6.8.2
streamlit==1.36.0
tenacity==8.4.2
text-unidecode==1.3
toml==0.10.2
tomli_w==1.0.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
xhs==0.2.13
yagmail==0.15.293
yarl==1.9.4
ykdl==1.8.2
yt-dlp==2024.7.2
